+++ Assets/Translations/Editor/Strings.cs
@@ -1,8 +1,10 @@
 namespace Ettmetal.Translation.Editor {
-    internal static class Strings {
-        public static readonly string NoStrings = "No strings to edit";
+	internal static class Strings {
+		public static readonly string NoStrings = "No strings to edit";
 		public static readonly string SelectedKeyNotFound = "The localisation key currently assigned to this Component was not found in any locale.";
 		public static readonly string NoKeys = "No localisation keys found";
 		public static readonly string ResourcesRoot = "Assets/Resources/";
-    }
+		public static readonly string TextPrefabPath = "Prefabs/Localized Text";
+		public static readonly string DropdownPrefabPath = "Prefabs/Locale Dropdown";
+	}
 }